Rename WMMA triple to RDNA3 and move cooperative matrix config to KernelConfig
