+++ compiler/src/iree/compiler/Codegen/SPIRV/test/BUILD
@@ -22,6 +22,7 @@ iree_lit_test_suite(
             "config_adreno_conv.mlir",
             "config_adreno_matmul.mlir",
             "config_amd_matmul.mlir",
+            "config_amd_matmul_cooperative_ops.mlir",
             "config_default_conv.mlir",
             "config_default_linalg_ext_ops.mlir",
             "config_default_linalg_ops.mlir",
